# docs/swagger.yaml
definitions:
...
  dokumen.SiupModel:
    properties:
      file:
        type: string
      id:
        type: string
      id_bumd:
        type: string
      instansi_pemberi:
        type: string
      is_seumur_hidup:
        type: integer
      klasifikasi:
        type: string
      kualifikasi:
        type: integer
      masa_berlaku:
        type: string
      nomor:
        type: string
      tanggal:
        type: string
    type: object
...
  utils.DataValidationError:
    properties:
      field:
        example: email
        type: string
      message:
        example: Invalid email address
        type: string
    type: object
...
  utils.RequestError:
    properties:
      code:
        example: 422
        type: integer
      fields:
        items:
          $ref: '#/definitions/utils.DataValidationError'
        type: array
      message:
        example: Invalid email address
        type: string
    type: object
info:
  contact:
    email: lifelinejar@mail.com
    name: API Support
  description: BUMD Service Rest API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BUMD Service
  version: "1.0"
paths:
...
  /strict/bumd/{id_bumd}/siup:
    get:
      description: get data siup.
      operationId: siup-index
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      - description: Kualifikasi, 1 = Kecil, 2 = Non Kecil
        in: query
        name: kualifikasi
        type: integer
      - description: Is Seumur Hidup
        in: query
        name: is_seumur_hidup
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/dokumen.SiupModel'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: get data siup
      tags:
      - Siup
    post:
      consumes:
      - multipart/form-data
      description: create data siup.
      operationId: siup-create
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Nomor
        in: formData
        name: nomor
        type: string
      - description: Instansi Pemberi
        in: formData
        name: instansi_pemberi
        type: string
      - description: Tanggal
        in: formData
        name: tanggal
        type: string
      - description: Klasifikasi
        in: formData
        name: klasifikasi
        type: string
      - description: Kualifikasi
        in: formData
        name: kualifikasi
        type: integer
      - description: Masa Berlaku
        in: formData
        name: masa_berlaku
        type: string
      - description: File
        in: formData
        name: file
        type: file
      responses:
        "200":
          description: Success
          schema:
            type: boolean
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: create data siup
      tags:
      - Siup
  /strict/bumd/{id_bumd}/siup/{id}:
    delete:
      consumes:
      - application/json
      description: delete data siup.
      operationId: siup-delete
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Id SIUP
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success
          schema:
            type: boolean
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: delete data siup
      tags:
      - Siup
    get:
      description: get data siup.
      operationId: siup-view
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Id SIUP
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dokumen.SiupModel'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: get data siup
      tags:
      - Siup
    put:
      consumes:
      - multipart/form-data
      description: update data siup.
      operationId: siup-update
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Id SIUP
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Nomor
        in: formData
        name: nomor
        required: true
        type: string
      - description: Instansi Pemberi
        in: formData
        name: instansi_pemberi
        required: true
        type: string
      - description: Tanggal
        in: formData
        name: tanggal
        required: true
        type: string
      - description: Klasifikasi
        in: formData
        name: klasifikasi
        required: true
        type: string
      - description: Kualifikasi
        in: formData
        name: kualifikasi
        required: true
        type: integer
      - description: Masa Berlaku
        in: formData
        name: masa_berlaku
        type: string
      - description: File
        in: formData
        name: file
        type: file
      responses:
        "200":
          description: Success
          schema:
            type: boolean
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: update data siup
      tags:
      - Siup
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"